3-to_str.py

+++ 0x00-python_variable_annotations/3-to_str.py
@@ -2,6 +2,6 @@
 """ to_str function """
 
 
-def to_str(a: int) -> str:
+def to_str(a: float) -> str:
     """ return from int to string """
     return str(a)
